A few updates.
Added a neutron density attribute.
For now, this is a scalar value representing the energy-intrated neutron density.
This will later become a vector of group-wise densities for implementation of the neutron material method.

The initial conditions were also halved to be consistent with an isotropic angular flux. This would be division by 4pi if using azimuthal angles.